\name{gfmR-package}
\alias{gfmR-package}
\alias{gfmR}
\docType{package}
\title{
\packageTitle{gfmR}
}
\description{
\packageDescription{gfmR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gfmR}
\packageIndices{gfmR}
The most improtant functions are GroupFusedMulti and GFMR.cv which implement methodology
proposed by the manuscript "Automatic Response Category Combination in Multinomial Logistic Regression" by Brad Price, Charles Geyer, and Adam Rothman.  
}
\author{
\packageAuthor{Brad Price}

Maintainer: \packageMaintainer{Brad Price brad.price@mail.wvu.edu}
}
\references{
Price, B.S, Geyer, C.J. and Rothman, A.J. "Automatic Response Category Combination in Multinomial Logistic Regression."  https://arxiv.org/abs/1705.03594. 
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package
models
regression
multivariate
}

\examples{
~~ simple examples of the most important functions ~~
}
